+++ src/coalesce-vue-vuetify3/src/components/display/c-loader-status.vue
@@ -341,7 +341,10 @@ const showContent = computed(() => {
       // initial content is off, and either:
       !flags["initial-content"] &&
       // loader has not yet loaded
-      !loader._hasLoaded.value
+      (loader.wasSuccessful == null ||
+        // or loader has loaded, but it errored and there's no current result
+        // (implying it has never successfully loaded).
+        (loader.wasSuccessful === false && !loader.hasResult))
     ) {
       return false;
     }